+++ test/jest/__mock__/index.js
@@ -5,3 +5,4 @@ import './stripesIcon.mock';
 import './stripesComponents.mock';
 import './finalFormComponents.mock';
 import './stripesSmartComponents.mock';
+import './stripesCore.mock';